+++ docs/conf.py
@@ -30,6 +30,8 @@ extensions = [
     'nbsphinx_link',
 ]
 
+default_role = 'py:obj'
+
 # -- sphinx.ext.autodoc
 autoclass_content = 'both'  # class and __init__ docstrings are concatenated
 autodoc_default_flags = ['members']