+++ CHANGES.rst
@@ -30,4 +30,4 @@ Release history
 Initial release of Nengo Extras!
 Tested with Nengo 2.7.0, but should work with earlier versions.
 If you run into any issues, please
-`file a bug report <https://github.com/nengo/nengo_extras/issues/new>`_.
+`file a bug report <https://github.com/nengo/nengo-extras/issues/new>`_.